+++ tests/Table/FactoryTest.php
@@ -52,7 +52,7 @@ class FactoryTest extends \PHPUnit_Framework_TestCase
     protected function getContainer()
     {
         $app  = new Container();
-        $app['Illuminate\Contracts\Config\Repository'] = $config = m::mock('\Illuminate\Contracts\Config\Repository');
+        $app['Illuminate\Contracts\Config\Repository'] = $config = m::mock('\Illuminate\Contracts\Config\Repository, \Orchestra\Contracts\Config\PackageRepository');
         $app['request'] = m::mock('\Illuminate\Http\Request');
         $app['translator'] = m::mock('\Illuminate\Translation\Translator');
         $app['view'] = m::mock('\Illuminate\Contracts\View\Factory');